fix dab-detr inference bugs
