+++ projects/dab_detr/configs/models/dab_detr_r50.py
@@ -1,8 +1,15 @@
+import torch.nn as nn
+
 from ideadet.modeling.utils import Joiner, MaskedBackbone
 from ideadet.modeling.matcher import DabMatcher
 from ideadet.modeling.criterion import DabCriterion
-from ideadet.layers.position_embedding import (
+from ideadet.layers import (
+    MultiheadAttention,
+    ConditionalSelfAttention,
+    ConditionalCrossAttention,
     PositionEmbeddingSine,
+    FFN,
+    BaseTransformerLayer,
 )
 
 from detectron2.modeling.backbone import ResNet, BasicStem
@@ -10,7 +17,9 @@ from detectron2.config import LazyCall as L
 
 from modeling import (
     DABDETR,
-    Transformer,
+    DabDetrTransformer,
+    DabDetrTransformerDecoder,
+    DabDetrTransformerEncoder,
 )
 
 
@@ -32,16 +41,60 @@ model = L(DABDETR)(
             num_pos_feats=128, temperature=20, normalize=True
         ),
     ),
-    transformer=L(Transformer)(
-        d_model=256,
-        dropout=0.1,
-        nhead=8,
-        dim_feedforward=2048,
-        num_encoder_layers=6,
-        num_decoder_layers=6,
-        normalize_before=False,
-        activation="prelu",
-        return_intermediate_dec="${..aux_loss}",
+    transformer=L(DabDetrTransformer)(
+        encoder=L(DabDetrTransformerEncoder)(
+            transformer_layers=L(BaseTransformerLayer)(
+                attn=L(MultiheadAttention)(
+                    embed_dim=256,
+                    num_heads=8,
+                    attn_drop=0.1,
+                    batch_first=False,
+                ),
+                ffn=L(FFN)(
+                    embed_dim=256,
+                    feedforward_dim=2048,
+                    ffn_drop=0.1,
+                    activation=L(nn.PReLU)(),
+                ),
+                norm=L(nn.LayerNorm)(normalized_shape=256),
+                operation_order=("self_attn", "norm", "ffn", "norm"),
+            ),
+            num_layers=6,
+            post_norm=False,
+        ),
+        decoder=L(DabDetrTransformerDecoder)(
+            num_layers=6,
+            return_intermediate=True,
+            query_dim=4,
+            modulate_hw_attn=True,
+            post_norm=True,
+            transformer_layers=L(BaseTransformerLayer)(
+                attn=[
+                    L(ConditionalSelfAttention)(
+                        embed_dim=256,
+                        num_heads=8,
+                        attn_drop=0.1,
+                        batch_first=False,
+                    ),
+                    L(ConditionalCrossAttention)(
+                        embed_dim=256,
+                        num_heads=8,
+                        attn_drop=0.1,
+                        batch_first=False,
+                    ),
+                ],
+                ffn=L(FFN)(
+                    embed_dim=256,
+                    feedforward_dim=2048,
+                    ffn_drop=0.1,
+                    activation=L(nn.PReLU)(),
+                ),
+                norm=L(nn.LayerNorm)(
+                    normalized_shape=256,
+                ),
+                operation_order=("self_attn", "norm", "cross_attn", "norm", "ffn", "norm"),
+            ),
+        ),
     ),
     num_classes=80,
     num_queries=300,
@@ -73,7 +126,7 @@ model = L(DABDETR)(
 if model.aux_loss:
     weight_dict = model.criterion.weight_dict
     aux_weight_dict = {}
-    for i in range(model.transformer.num_decoder_layers - 1):
+    for i in range(model.transformer.decoder.num_layers - 1):
         aux_weight_dict.update({k + f"_{i}": v for k, v in weight_dict.items()})
     weight_dict.update(aux_weight_dict)
     model.criterion.weight_dict = weight_dict
