+++ projects/dab_detr/modeling/__init__.py
@@ -1,6 +1,5 @@
 from .dab_detr import DABDETR
-from .dab_transformer import Transformer
-from .rebuild_transformer import (
+from .dab_transformer import (
     DabDetrTransformerEncoder,
     DabDetrTransformerDecoder,
     DabDetrTransformer,
