+++ projects/dab_detr/modeling/dab_transformer.py
@@ -1,305 +1,189 @@
-# ------------------------------------------------------------------------
-# DAB-DETR
-# Copyright (c) 2022 IDEA. All Rights Reserved.
-# Licensed under the Apache License, Version 2.0 [see LICENSE for details]
-# ------------------------------------------------------------------------
-# Modified from Conditional DETR (https://github.com/Atten4Vis/ConditionalDETR)
-# Copyright (c) 2021 Microsoft. All Rights Reserved.
-# Licensed under the Apache License, Version 2.0 [see LICENSE for details]
-# ------------------------------------------------------------------------
-# Modified from DETR (https://github.com/facebookresearch/detr)
-# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
-# ------------------------------------------------------------------------
-
-
-import copy
-from typing import Optional
+# coding=utf-8
+# Copyright 2022 The IDEA Authors. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
-from torch import Tensor
 
-from ideadet.layers import (
-    ConditionalCrossAttention,
-    ConditionalSelfAttention,
-    MLP,
-    get_sine_pos_embed,
-)
+from ideadet.layers import MLP, BaseTransformerLayer, TransformerLayerSequence, get_sine_pos_embed
 from ideadet.utils.misc import inverse_sigmoid
 
 
-class Transformer(nn.Module):
+class DabDetrTransformerEncoder(TransformerLayerSequence):
     def __init__(
         self,
-        d_model=512,
-        nhead=8,
-        num_queries=300,
-        num_encoder_layers=6,
-        num_decoder_layers=6,
-        dim_feedforward=2048,
-        dropout=0.1,
-        activation="relu",
-        normalize_before=False,
-        return_intermediate_dec=False,
-        query_dim=4,
-        keep_query_pos=False,
-        query_scale_type="cond_elewise",
-        num_patterns=0,
-        modulate_hw_attn=True,
-        bbox_embed_diff_each_layer=False,
+        transformer_layers: BaseTransformerLayer = None,
+        post_norm: bool = True,
+        num_layers: int = None,
     ):
-
-        super().__init__()
-
-        encoder_layer = TransformerEncoderLayer(
-            d_model, nhead, dim_feedforward, dropout, activation, normalize_before
+        super(DabDetrTransformerEncoder, self).__init__(
+            transformer_layers=transformer_layers, num_layers=num_layers
         )
-        encoder_norm = nn.LayerNorm(d_model) if normalize_before else None
-        self.encoder = TransformerEncoder(encoder_layer, num_encoder_layers, encoder_norm)
+        self.embed_dim = self.layers[0].embed_dim
+        self.pre_norm = self.layers[0].pre_norm
+        self.query_scale = MLP(self.embed_dim, self.embed_dim, self.embed_dim, 2)
 
-        decoder_layer = TransformerDecoderLayer(
-            d_model,
-            nhead,
-            dim_feedforward,
-            dropout,
-            activation,
-            normalize_before,
-            keep_query_pos=keep_query_pos,
-        )
-        decoder_norm = nn.LayerNorm(d_model)
-        self.decoder = TransformerDecoder(
-            decoder_layer,
-            num_decoder_layers,
-            decoder_norm,
-            return_intermediate=return_intermediate_dec,
-            d_model=d_model,
-            query_dim=query_dim,
-            keep_query_pos=keep_query_pos,
-            query_scale_type=query_scale_type,
-            modulate_hw_attn=modulate_hw_attn,
-            bbox_embed_diff_each_layer=bbox_embed_diff_each_layer,
-        )
-
-        self._reset_parameters()
-        assert query_scale_type in ["cond_elewise", "cond_scalar", "fix_elewise"]
-
-        self.d_model = d_model
-        self.nhead = nhead
-        self.dec_layers = num_decoder_layers
-        self.num_queries = num_queries
-        self.num_patterns = num_patterns
-        if not isinstance(num_patterns, int):
-            Warning("num_patterns should be int but {}".format(type(num_patterns)))
-            self.num_patterns = 0
-        if self.num_patterns > 0:
-            self.patterns = nn.Embedding(self.num_patterns, d_model)
-
-    def _reset_parameters(self):
-        for p in self.parameters():
-            if p.dim() > 1:
-                nn.init.xavier_uniform_(p)
-
-    def forward(self, src, mask, refpoint_embed, pos_embed):
-        # flatten NxCxHxW to HWxNxC
-        bs, c, h, w = src.shape
-        src = src.flatten(2).permute(2, 0, 1)
-        pos_embed = pos_embed.flatten(2).permute(2, 0, 1)
-        refpoint_embed = refpoint_embed.unsqueeze(1).repeat(1, bs, 1)
-        mask = mask.flatten(1)
-        memory = self.encoder(src, src_key_padding_mask=mask, pos=pos_embed)
-
-        # query_embed = gen_sineembed_for_position(refpoint_embed)
-        num_queries = refpoint_embed.shape[0]
-        if self.num_patterns == 0:
-            tgt = torch.zeros(num_queries, bs, self.d_model, device=refpoint_embed.device)
+        if post_norm:
+            self.post_norm_layer = nn.LayerNorm(self.embed_dim)
         else:
-            tgt = (
-                self.patterns.weight[:, None, None, :]
-                .repeat(1, self.num_queries, bs, 1)
-                .flatten(0, 1)
-            )  # n_q*n_pat, bs, d_model
-            refpoint_embed = refpoint_embed.repeat(
-                self.num_patterns, 1, 1
-            )  # n_q*n_pat, bs, d_model
-            # import ipdb; ipdb.set_trace()
-        hs, references = self.decoder(
-            tgt,
-            memory,
-            memory_key_padding_mask=mask,
-            pos=pos_embed,
-            refpoints_unsigmoid=refpoint_embed,
-        )
-        return hs, references
-
-
-class TransformerEncoder(nn.Module):
-    def __init__(self, encoder_layer, num_layers, norm=None, d_model=256):
-        super().__init__()
-        self.layers = _get_clones(encoder_layer, num_layers)
-        self.num_layers = num_layers
-        self.query_scale = MLP(d_model, d_model, d_model, 2)
-        self.norm = norm
+            self.post_norm_layer = None
 
     def forward(
         self,
-        src,
-        mask: Optional[Tensor] = None,
-        src_key_padding_mask: Optional[Tensor] = None,
-        pos: Optional[Tensor] = None,
+        query,
+        key,
+        value,
+        query_pos=None,
+        key_pos=None,
+        attn_masks=None,
+        query_key_padding_mask=None,
+        key_padding_mask=None,
+        **kwargs,
     ):
-        output = src
 
-        for layer_id, layer in enumerate(self.layers):
-            # rescale the content and pos sim
-            pos_scales = self.query_scale(output)
-            output = layer(
-                output,
-                src_mask=mask,
-                src_key_padding_mask=src_key_padding_mask,
-                pos=pos * pos_scales,
+        for layer in self.layers:
+            position_scales = self.query_scale(query)
+            query = layer(
+                query,
+                key,
+                value,
+                query_pos=query_pos * position_scales,
+                attn_masks=attn_masks,
+                query_key_padding_mask=query_key_padding_mask,
+                key_padding_mask=key_padding_mask,
+                **kwargs,
             )
 
-        if self.norm is not None:
-            output = self.norm(output)
+        if self.post_norm_layer is not None:
+            query = self.post_norm_layer(query)
+        return query
 
-        return output
 
-
-class TransformerDecoder(nn.Module):
+class DabDetrTransformerDecoder(TransformerLayerSequence):
     def __init__(
         self,
-        decoder_layer,
-        num_layers,
-        norm=None,
-        return_intermediate=False,
-        d_model=256,
-        query_dim=2,
-        keep_query_pos=False,
-        query_scale_type="cond_elewise",
-        modulate_hw_attn=True,
-        bbox_embed_diff_each_layer=False,
+        transformer_layers: BaseTransformerLayer = None,
+        num_layers: int = None,
+        query_dim: int = 4,
+        modulate_hw_attn: bool = True,
+        post_norm: bool = True,
+        return_intermediate: bool = True,
     ):
-        super().__init__()
-        self.layers = _get_clones(decoder_layer, num_layers)
-        self.num_layers = num_layers
-        self.norm = norm
+        super().__init__(transformer_layers, num_layers)
         self.return_intermediate = return_intermediate
-        assert return_intermediate
-        self.query_dim = query_dim
+        self.embed_dim = self.layers[0].embed_dim
 
-        assert query_scale_type in ["cond_elewise", "cond_scalar", "fix_elewise"]
-        self.query_scale_type = query_scale_type
-        if query_scale_type == "cond_elewise":
-            self.query_scale = MLP(d_model, d_model, d_model, 2)
-        elif query_scale_type == "cond_scalar":
-            self.query_scale = MLP(d_model, d_model, 1, 2)
-        elif query_scale_type == "fix_elewise":
-            self.query_scale = nn.Embedding(num_layers, d_model)
-        else:
-            raise NotImplementedError("Unknown query_scale_type: {}".format(query_scale_type))
-
-        self.ref_point_head = MLP(query_dim // 2 * d_model, d_model, d_model, 2)
+        self.query_scale = MLP(self.embed_dim, self.embed_dim, self.embed_dim, 2)
+        self.ref_point_head = MLP(
+            query_dim // 2 * self.embed_dim, self.embed_dim, self.embed_dim, 2
+        )
 
         self.bbox_embed = None
-        self.d_model = d_model
-        self.modulate_hw_attn = modulate_hw_attn
-        self.bbox_embed_diff_each_layer = bbox_embed_diff_each_layer
-
         if modulate_hw_attn:
-            self.ref_anchor_head = MLP(d_model, d_model, 2, 2)
+            self.ref_anchor_head = MLP(self.embed_dim, self.embed_dim, 2, 2)
+        self.modulate_hw_attn = modulate_hw_attn
 
-        if not keep_query_pos:
-            for layer_id in range(num_layers - 1):
-                self.layers[layer_id + 1].cross_attn.query_pos_proj = None
+        if post_norm:
+            self.post_norm_layer = nn.LayerNorm(self.embed_dim)
+        else:
+            self.post_norm_layer = None
+
+        for idx in range(num_layers - 1):
+            self.layers[idx + 1].attentions[1].query_pos_proj = None
 
     def forward(
         self,
-        tgt,
-        memory,
-        tgt_mask: Optional[Tensor] = None,
-        memory_mask: Optional[Tensor] = None,
-        tgt_key_padding_mask: Optional[Tensor] = None,
-        memory_key_padding_mask: Optional[Tensor] = None,
-        pos: Optional[Tensor] = None,
-        refpoints_unsigmoid: Optional[Tensor] = None,  # num_queries, bs, 2
+        query,
+        key,
+        value,
+        query_pos=None,
+        key_pos=None,
+        attn_masks=None,
+        query_key_padding_mask=None,
+        key_padding_mask=None,
+        refpoints_embed=None,
+        **kwargs,
     ):
-        output = tgt
-
         intermediate = []
-        reference_points = refpoints_unsigmoid.sigmoid()
-        ref_points = [reference_points]
 
-        # import ipdb; ipdb.set_trace()
+        reference_points = refpoints_embed.sigmoid()
+        refpoints = [reference_points]
 
-        for layer_id, layer in enumerate(self.layers):
-            obj_center = reference_points[..., : self.query_dim]  # [num_queries, batch_size, 2]
-            # get sine embedding for the query vector
+        for idx, layer in enumerate(self.layers):
+            obj_center = reference_points[..., : self.embed_dim]
             query_sine_embed = get_sine_pos_embed(obj_center)
             query_pos = self.ref_point_head(query_sine_embed)
 
-            # For the first decoder layer, we do not apply transformation over p_s
-            if self.query_scale_type != "fix_elewise":
-                if layer_id == 0:
-                    pos_transformation = 1
-                else:
-                    pos_transformation = self.query_scale(output)
+            # do not apply transform in position in the first decoder layer
+            if idx == 0:
+                position_transform = 1
             else:
-                pos_transformation = self.query_scale.weight[layer_id]
+                position_transform = self.query_scale(query)
 
-            # apply transformation
-            query_sine_embed = query_sine_embed[..., : self.d_model] * pos_transformation
+            # apply position transform
+            query_sine_embed = query_sine_embed[..., : self.embed_dim] * position_transform
 
-            # modulated HW attentions
             if self.modulate_hw_attn:
-                refHW_cond = self.ref_anchor_head(output).sigmoid()  # nq, bs, 2
-                query_sine_embed[..., self.d_model // 2 :] *= (
-                    refHW_cond[..., 0] / obj_center[..., 2]
+                ref_hw_cond = self.ref_anchor_head(query).sigmoid()
+                query_sine_embed[..., self.embed_dim // 2 :] *= (
+                    ref_hw_cond[..., 0] / obj_center[..., 2]
                 ).unsqueeze(-1)
-                query_sine_embed[..., : self.d_model // 2] *= (
-                    refHW_cond[..., 1] / obj_center[..., 3]
+                query_sine_embed[..., : self.embed_dim // 2] *= (
+                    ref_hw_cond[..., 1] / obj_center[..., 3]
                 ).unsqueeze(-1)
 
-            output = layer(
-                output,
-                memory,
-                tgt_mask=tgt_mask,
-                memory_mask=memory_mask,
-                tgt_key_padding_mask=tgt_key_padding_mask,
-                memory_key_padding_mask=memory_key_padding_mask,
-                pos=pos,
+            query = layer(
+                query,
+                key,
+                value,
                 query_pos=query_pos,
+                key_pos=key_pos,
                 query_sine_embed=query_sine_embed,
-                is_first=(layer_id == 0),
+                attn_masks=attn_masks,
+                query_key_padding_mask=query_key_padding_mask,
+                key_padding_mask=key_padding_mask,
+                is_first_layer=(idx == 0),
+                **kwargs,
             )
 
             # iter update
             if self.bbox_embed is not None:
-                if self.bbox_embed_diff_each_layer:
-                    tmp = self.bbox_embed[layer_id](output)
-                else:
-                    tmp = self.bbox_embed(output)
-                # import ipdb; ipdb.set_trace()
-                tmp[..., : self.query_dim] += inverse_sigmoid(reference_points)
-                new_reference_points = tmp[..., : self.query_dim].sigmoid()
-                if layer_id != self.num_layers - 1:
-                    ref_points.append(new_reference_points)
+                temp = self.bbox_embed(query)
+                temp[..., : self.embed_dim] += inverse_sigmoid(reference_points)
+                new_reference_points = temp[..., : self.embed_dim].sigmoid()
+
+                if idx != self.num_layers - 1:
+                    refpoints.append(new_reference_points)
                 reference_points = new_reference_points.detach()
 
             if self.return_intermediate:
-                intermediate.append(self.norm(output))
+                if self.post_norm_layer is not None:
+                    intermediate.append(self.post_norm_layer(query))
+                else:
+                    intermediate.append(query)
 
-        if self.norm is not None:
-            output = self.norm(output)
+        if self.post_norm_layer is not None:
+            query = self.post_norm_layer(query)
             if self.return_intermediate:
                 intermediate.pop()
-                intermediate.append(output)
+                intermediate.append(query)
 
         if self.return_intermediate:
             if self.bbox_embed is not None:
                 return [
                     torch.stack(intermediate).transpose(1, 2),
-                    torch.stack(ref_points).transpose(1, 2),
+                    torch.stack(refpoints).transpose(1, 2),
                 ]
             else:
                 return [
@@ -307,186 +191,45 @@ class TransformerDecoder(nn.Module):
                     reference_points.unsqueeze(0).transpose(1, 2),
                 ]
 
-        return output.unsqueeze(0)
+        return query.unsqueeze(0)
 
 
-class TransformerEncoderLayer(nn.Module):
-    def __init__(
-        self,
-        d_model,
-        nhead,
-        dim_feedforward=2048,
-        dropout=0.1,
-        activation="relu",
-        normalize_before=False,
-    ):
-        super().__init__()
-        self.self_attn = nn.MultiheadAttention(d_model, nhead, dropout=dropout)
-        # Implementation of Feedforward model
-        self.linear1 = nn.Linear(d_model, dim_feedforward)
-        self.dropout = nn.Dropout(dropout)
-        self.linear2 = nn.Linear(dim_feedforward, d_model)
+class DabDetrTransformer(nn.Module):
+    def __init__(self, encoder=None, decoder=None):
+        super(DabDetrTransformer, self).__init__()
+        self.encoder = encoder
+        self.decoder = decoder
+        self.embed_dim = self.encoder.embed_dim
 
-        self.norm1 = nn.LayerNorm(d_model)
-        self.norm2 = nn.LayerNorm(d_model)
-        self.dropout1 = nn.Dropout(dropout)
-        self.dropout2 = nn.Dropout(dropout)
+        self.init_weights()
 
-        self.activation = _get_activation_fn(activation)
-        self.normalize_before = normalize_before
+    def init_weights(self):
+        for p in self.parameters():
+            if p.dim() > 1:
+                nn.init.xavier_uniform_(p)
 
-    def with_pos_embed(self, tensor, pos: Optional[Tensor]):
-        return tensor if pos is None else tensor + pos
+    def forward(self, x, mask, refpoints_embed, pos_embed):
+        bs, c, h, w = x.shape
+        x = x.view(bs, c, -1).permute(2, 0, 1)
+        pos_embed = pos_embed.view(bs, c, -1).permute(2, 0, 1)
+        refpoints_embed = refpoints_embed.unsqueeze(1).repeat(1, bs, 1)
+        mask = mask.view(bs, -1)
+        memory = self.encoder(
+            query=x,
+            key=None,
+            value=None,
+            query_pos=pos_embed,
+            query_key_padding_mask=mask,
+        )
+        num_queries = refpoints_embed.shape[0]
+        target = torch.zeros(num_queries, bs, self.embed_dim, device=refpoints_embed.device)
 
-    def forward(
-        self,
-        src,
-        src_mask: Optional[Tensor] = None,
-        src_key_padding_mask: Optional[Tensor] = None,
-        pos: Optional[Tensor] = None,
-    ):
-        q = k = self.with_pos_embed(src, pos)
-        src2 = self.self_attn(
-            q, k, value=src, attn_mask=src_mask, key_padding_mask=src_key_padding_mask
-        )[0]
-        src = src + self.dropout1(src2)
-        src = self.norm1(src)
-        src2 = self.linear2(self.dropout(self.activation(self.linear1(src))))
-        src = src + self.dropout2(src2)
-        src = self.norm2(src)
-        return src
-
-
-class TransformerDecoderLayer(nn.Module):
-    def __init__(
-        self,
-        d_model,
-        nhead,
-        dim_feedforward=2048,
-        dropout=0.1,
-        activation="relu",
-        normalize_before=False,
-        keep_query_pos=False,
-        rm_self_attn_decoder=False,
-    ):
-        super().__init__()
-        # Decoder Self-Attention
-        if not rm_self_attn_decoder:
-            self.self_attn = ConditionalSelfAttention(d_model, nhead, dropout)
-
-            self.norm1 = nn.LayerNorm(d_model)
-            self.dropout1 = nn.Dropout(dropout)
-
-        # Decoder Cross-Attention
-        self.cross_attn = ConditionalCrossAttention(d_model, nhead, attn_drop=dropout, vdim=d_model)
-
-        self.nhead = nhead
-        self.rm_self_attn_decoder = rm_self_attn_decoder
-
-        # Implementation of Feedforward model
-        self.linear1 = nn.Linear(d_model, dim_feedforward)
-        self.dropout = nn.Dropout(dropout)
-        self.linear2 = nn.Linear(dim_feedforward, d_model)
-
-        self.norm2 = nn.LayerNorm(d_model)
-        self.norm3 = nn.LayerNorm(d_model)
-        self.dropout2 = nn.Dropout(dropout)
-        self.dropout3 = nn.Dropout(dropout)
-
-        self.activation = _get_activation_fn(activation)
-        self.normalize_before = normalize_before
-
-    def with_pos_embed(self, tensor, pos: Optional[Tensor]):
-        return tensor if pos is None else tensor + pos
-
-    def forward(
-        self,
-        tgt,
-        memory,
-        tgt_mask: Optional[Tensor] = None,
-        memory_mask: Optional[Tensor] = None,
-        tgt_key_padding_mask: Optional[Tensor] = None,
-        memory_key_padding_mask: Optional[Tensor] = None,
-        pos: Optional[Tensor] = None,
-        query_pos: Optional[Tensor] = None,
-        query_sine_embed=None,
-        is_first=False,
-    ):
-
-        # ========== Begin of Self-Attention =============
-        if not self.rm_self_attn_decoder:
-            # Apply projections here
-            # shape: num_queries x batch_size x 256
-
-            tgt2 = self.self_attn(
-                tgt,
-                tgt,
-                value=tgt,
-                attn_mask=tgt_mask,
-                key_padding_mask=tgt_key_padding_mask,
-                query_pos=query_pos,
-                key_pos=query_pos,
-            )
-            # ========== End of Self-Attention =============
-
-            # tgt = tgt + self.dropout1(tgt2)
-            tgt = self.norm1(tgt2)
-
-        # ========== Begin of Cross-Attention =============
-        # Apply projections here
-        # shape: num_queries x batch_size x 256
-        tgt2 = self.cross_attn(
-            query=tgt,
+        hidden_state, references = self.decoder(
+            query=target,
             key=memory,
             value=memory,
-            attn_mask=memory_mask,
-            key_padding_mask=memory_key_padding_mask,
-            query_sine_embed=query_sine_embed,
-            query_pos=query_pos,
-            key_pos=pos,
-            is_first_layer=is_first,
+            key_pos=pos_embed,
+            refpoints_embed=refpoints_embed,
         )
-        # ========== End of Cross-Attention =============
-        # tgt2 = tgt + tgt2
-        # tgt = tgt + self.dropout2(tgt2)
-        tgt = self.norm2(tgt2)
-        tgt2 = self.linear2(self.dropout(self.activation(self.linear1(tgt))))
-        tgt = tgt + self.dropout3(tgt2)
-        tgt = self.norm3(tgt)
-        return tgt
 
-
-def _get_clones(module, N):
-    return nn.ModuleList([copy.deepcopy(module) for i in range(N)])
-
-
-def build_transformer(args):
-    return Transformer(
-        d_model=args.hidden_dim,
-        dropout=args.dropout,
-        nhead=args.nheads,
-        num_queries=args.num_queries,
-        dim_feedforward=args.dim_feedforward,
-        num_encoder_layers=args.enc_layers,
-        num_decoder_layers=args.dec_layers,
-        normalize_before=args.pre_norm,
-        return_intermediate_dec=True,
-        query_dim=4,
-        activation=args.transformer_activation,
-        num_patterns=args.num_patterns,
-    )
-
-
-def _get_activation_fn(activation):
-    """Return an activation function given a string"""
-    if activation == "relu":
-        return F.relu
-    if activation == "gelu":
-        return F.gelu
-    if activation == "glu":
-        return F.glu
-    if activation == "prelu":
-        return nn.PReLU()
-    if activation == "selu":
-        return F.selu
-    raise RuntimeError(f"activation should be relu/gelu, not {activation}.")
+        return hidden_state, references
