--- projects/dab_detr/modeling/rebuild_transformer.py
@@ -1,238 +1,0 @@
-# coding=utf-8
-# Copyright 2022 The IDEA Authors. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import torch
-import torch.nn as nn
-
-from ideadet.layers import (
-    BaseTransformerLayer, 
-    TransformerLayerSequence,
-    MLP,
-    get_sine_pos_embed,
-)
-from ideadet.utils.misc import inverse_sigmoid
-
-
-class DabDetrTransformerEncoder(TransformerLayerSequence):
-    def __init__(
-        self,
-        transformer_layers: BaseTransformerLayer = None,
-        post_norm: bool = True,
-        num_layers: int = None,
-    ):
-        super(DabDetrTransformerEncoder, self).__init__(
-            transformer_layers=transformer_layers, num_layers=num_layers
-        )
-        self.embed_dim = self.layers[0].embed_dim
-        self.pre_norm = self.layers[0].pre_norm
-        self.query_scale = MLP(self.embed_dim, self.embed_dim, self.embed_dim, 2)
-
-        if post_norm:
-            self.post_norm_layer = nn.LayerNorm(self.embed_dim)
-        else:
-            self.post_norm_layer = None
-
-    def forward(
-        self,
-        query,
-        key,
-        value,
-        query_pos=None,
-        key_pos=None,
-        attn_masks=None,
-        query_key_padding_mask=None,
-        key_padding_mask=None,
-        **kwargs,
-    ):
-
-        for layer in self.layers:
-            position_scales = self.query_scale(query)
-            query = layer(
-                query,
-                key,
-                value,
-                query_pos=query_pos * position_scales,
-                attn_masks=attn_masks,
-                query_key_padding_mask=query_key_padding_mask,
-                key_padding_mask=key_padding_mask,
-                **kwargs,
-            )
-
-        if self.post_norm_layer is not None:
-            query = self.post_norm_layer(query)
-        return query
-
-
-class DabDetrTransformerDecoder(TransformerLayerSequence):
-    def __init__(
-        self, 
-        transformer_layers: BaseTransformerLayer = None, 
-        num_layers: int = None,
-        query_dim: int = 4,
-        modulate_hw_attn: bool = True,
-        post_norm: bool = True,
-        return_intermediate: bool = True,
-    ):
-        super().__init__(transformer_layers, num_layers)
-        self.return_intermediate = return_intermediate
-        self.embed_dim = self.layers[0].embed_dim
-
-        self.query_scale = MLP(self.embed_dim, self.embed_dim, self.embed_dim, 2)
-        self.ref_point_head = MLP(query_dim // 2 * self.embed_dim, self.embed_dim, self.embed_dim, 2)
-
-        self.bbox_embed = None
-        if modulate_hw_attn:
-            self.ref_anchor_head = MLP(self.embed_dim, self.embed_dim, 2, 2)
-        self.modulate_hw_attn = modulate_hw_attn
-
-        if post_norm:
-            self.post_norm_layer = nn.LayerNorm(self.embed_dim)
-        else:
-            self.post_norm_layer = None
-
-        for idx in range(num_layers - 1):
-            self.layers[idx + 1].attentions[1].query_pos_proj = None
-
-
-    def forward(
-        self,
-        query,
-        key,
-        value,
-        query_pos=None,
-        key_pos=None,
-        attn_masks=None,
-        query_key_padding_mask=None,
-        key_padding_mask=None,
-        refpoints_embed=None,
-        **kwargs,
-    ):
-        intermediate = []
-
-        reference_points = refpoints_embed.sigmoid()
-        refpoints = [reference_points]
-
-        for idx, layer in enumerate(self.layers):
-            obj_center = reference_points[..., : self.embed_dim]
-            query_sine_embed = get_sine_pos_embed(obj_center)
-            query_pos = self.ref_point_head(query_sine_embed)
-
-            # do not apply transform in position in the first decoder layer
-            if idx == 0:
-                position_transform = 1
-            else:
-                position_transform = self.query_scale(query)
-            
-            # apply position transform
-            query_sine_embed = query_sine_embed[..., : self.embed_dim] * position_transform
-
-            if self.modulate_hw_attn:
-                ref_hw_cond = self.ref_anchor_head(query).sigmoid()
-                query_sine_embed[..., self.embed_dim // 2 :] *= (
-                    ref_hw_cond[..., 0] / obj_center[..., 2]
-                ).unsqueeze(-1)
-                query_sine_embed[..., : self.embed_dim // 2] *= (
-                    ref_hw_cond[..., 1] / obj_center[..., 3]
-                ).unsqueeze(-1)
-
-            query = layer(
-                query,
-                key,
-                value,
-                query_pos=query_pos,
-                key_pos=key_pos,
-                query_sine_embed=query_sine_embed,
-                attn_masks=attn_masks,
-                query_key_padding_mask=query_key_padding_mask,
-                key_padding_mask=key_padding_mask,
-                is_first=(idx==0),
-                **kwargs,
-            )
-
-            # iter update
-            if self.bbox_embed is not None:
-                temp = self.bbox_embed(query)
-                temp[..., : self.embed_dim] += inverse_sigmoid(reference_points)
-                new_reference_points = temp[..., : self.embed_dim].sigmoid()
-
-                if idx != self.num_layers - 1:
-                    refpoints.append(new_reference_points)
-                reference_points = new_reference_points.detach()
-            
-            if self.return_intermediate:
-                if self.post_norm_layer is not None:
-                    intermediate.append(self.post_norm_layer(query))
-                else:
-                    intermediate.append(query)
-
-        if self.post_norm_layer is not None:
-            query = self.post_norm_layer(query)
-            if self.return_intermediate:
-                intermediate.pop()
-                intermediate.append(query)        
-            
-        if self.return_intermediate:
-            if self.bbox_embed is not None:
-                return [
-                    torch.stack(intermediate).transpose(1, 2),
-                    torch.stack(refpoints).transpose(1, 2),
-                ]
-            else:
-                return [
-                    torch.stack(intermediate).transpose(1, 2),
-                    reference_points.unsqueeze(0).transpose(1, 2),
-                ]
-
-        return query.unsqueeze(0)
-
-
-class DabDetrTransformer(nn.Module):
-    def __init__(self, encoder=None, decoder=None):
-        super(DabDetrTransformer, self).__init__()
-        self.encoder = encoder
-        self.decoder = decoder
-        self.embed_dim = self.encoder.embed_dim
-
-        self.init_weights()
-
-    def init_weights(self):
-        for p in self.parameters():
-            if p.dim() > 1:
-                nn.init.xavier_uniform_(p)
-
-
-    def forward(self, x, mask, refpoints_embed, pos_embed):
-        bs, c, h, w = x.shape
-        x =  x.view(bs, c, -1).permute(2, 0, 1)
-        pos_embed = pos_embed.view(bs, c, -1).permute(2, 0, 1)
-        refpoints_embed = refpoints_embed.unsqueeze(1).repeat(1, bs, 1)
-        mask = mask.view(bs, -1)
-        memory = self.encoder(
-            query=x,
-            key=None,
-            value=None,
-            query_pos=pos_embed,
-            query_key_padding_mask=mask,
-        )
-        num_queries = refpoints_embed.shape[0]
-        target = torch.zeros(num_queries, bs, self.embed_dim, device=refpoints_embed.device)
-        hidden_state, references = self.decoder(
-            query=target,
-            key=memory,
-            value=memory,
-            key_pos=pos_embed,
-            refpoints_embed=refpoints_embed
-        )
-        return hidden_state, references
